# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flock Energy - Urja Meter Ops API",
    "description": "\nA clean, documented REST API over the legacy **Urja Meter Ops** portal.\n\nThe portal has no API. This service logs in as a normal user, manages the session,\nreverse-engineers the portal's internal endpoints, normalises its inconsistent payloads,\nand exposes the result as typed JSON that another program can consume without ever\ntouching the portal.\n\n**What it adds over the portal**\n\n* Filtering, sorting and pagination the portal cannot do (make, status, phase, build,\n  any hierarchy code).\n* The network hierarchy, reconstructed from per-meter paths into a queryable tree.\n* Real consumption, derived by differencing the portal's cumulative registers.\n* Proximity search over meter coordinates.\n* An explicit data-quality report instead of silently tidied-away inconsistencies.\n\nSee `PROTOCOL.md` for how the portal actually works, and `README.md` for design decisions\nand trade-offs.\n",
    "contact": {
      "name": "Flock Energy take-home submission"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/system/snapshot": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Snapshot provenance",
        "description": "Age, size, source and refresh history of the local index. Lets a caller decide for itself whether the data is fresh enough for its purpose.",
        "operationId": "snapshot_info_api_v1_system_snapshot_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotInfo"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/snapshot/refresh": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Force a snapshot refresh",
        "description": "Rebuild the index immediately, ignoring the TTL.\n\nDeliberately unauthenticated in this build because the service has no auth model at all (see README, *What I intentionally left out*). It is idempotent and read-only with respect to the portal, but it does cost one upstream export, so it should sit behind auth and a rate limit before any real deployment.",
        "operationId": "refresh_snapshot_api_v1_system_snapshot_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotInfo"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system/session": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Portal session status",
        "description": "Our view of the upstream session, for debugging re-authentication behaviour.",
        "operationId": "session_status_api_v1_system_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Session Status Api V1 System Session Get"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data-quality": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Data-quality report",
        "description": "Every inconsistency detected in the upstream data, aggregated with counts and sample meter ids.\n\nThis exists because hiding upstream mess behind a tidy API is how downstream teams end up debugging the wrong system. It also doubles as drift detection: if the portal's data changes, these counts move.",
        "operationId": "data_quality_api_v1_data_quality_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityReport"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Estate summary",
        "description": "Aggregate counts across the snapshot - the overview the portal never renders.",
        "operationId": "stats_api_v1_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Stats Api V1 Stats Get"
                }
              }
            }
          },
          "422": {
            "description": "Request parameters failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The upstream portal returned an unusable response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The upstream portal is unavailable or rate limiting us.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "The upstream portal did not respond in time.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataQualityIssue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "description": "'warning' or 'error'."
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "affected_count": {
            "type": "integer",
            "title": "Affected Count"
          },
          "sample_meter_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sample Meter Ids"
          }
        },
        "type": "object",
        "required": [
          "code",
          "level",
          "message",
          "affected_count"
        ],
        "title": "DataQualityIssue"
      },
      "DataQualityReport": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "issue_count": {
            "type": "integer",
            "title": "Issue Count"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/DataQualityIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "meter_count",
          "issue_count",
          "issues"
        ],
        "title": "DataQualityReport"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable slug. Branch on this, not on prose."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable, safe to display. Never a stack trace."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional structured context, e.g. invalid fields or a hint."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Correlation id; matches the X-Request-ID response header."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "The body of an :class:`ApiError` as clients receive it."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope returned for **every** non-2xx response.\n\nDeclared so the OpenAPI document describes the real contract. Without it FastAPI\nadvertises its own ``HTTPValidationError`` for 422 and nothing at all for 404/5xx,\nwhich makes generated clients fail to deserialise every error they receive."
      },
...
      "SnapshotInfo": {
        "properties": {
          "built_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Built At"
          },
          "age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Seconds"
          },
          "ttl_seconds": {
            "type": "integer",
            "title": "Ttl Seconds"
          },
          "is_stale": {
            "type": "boolean",
            "title": "Is Stale"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "transformer_count": {
            "type": "integer",
            "title": "Transformer Count"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Which upstream path produced it: 'export' or 'search'."
          },
          "build_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Duration Ms"
          },
          "refresh_count": {
            "type": "integer",
            "title": "Refresh Count",
            "default": 0
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "type": "object",
        "required": [
          "built_at",
          "age_seconds",
          "ttl_seconds",
          "is_stale",
          "meter_count",
          "transformer_count",
          "source"
        ],
        "title": "SnapshotInfo",
        "description": "Provenance for the local index - lets a caller reason about staleness itself."
      },
...
    }
  },
...
}